feat(autofix): Pull autofix timeline items out
- wraps marked conversions in useMemo
- pulls Expanded state into each element so that it only re-renders the one item on expand
